backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
pydicom==2.4.3
tensorflow==2.13.0